Updated ReactionsView so that it does not recycle reactions and also renders distiguishable children. Updated Dialogs so that now it closes when a reaction is selected.
