+++ app/src/main/java/com/adityaamolbavadekar/messenger/utils/Errors.kt
@@ -1,0 +1,4 @@
+package com.adityaamolbavadekar.messenger.utils
+
+enum class Errors {
+}
