+++ app/src/main/java/com/adityaamolbavadekar/messenger/views/ReadMoreTextView.kt
@@ -1,76 +1,105 @@
-/*
- *    Copyright 2023 Aditya Bavadekar
- *
- *    Licensed under the Apache License, Version 2.0 (the "License");
- *    you may not use this file except in compliance with the License.
- *    You may obtain a copy of the License at
- *
- *        http://www.apache.org/licenses/LICENSE-2.0
- *
- *    Unless required by applicable law or agreed to in writing, software
- *    distributed under the License is distributed on an "AS IS" BASIS,
- *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *    See the License for the specific language governing permissions and
- *    limitations under the License.
- */
-
 package com.adityaamolbavadekar.messenger.views
 
-import android.text.SpannableStringBuilder
-import android.text.Spanned
+import android.content.Context
+import android.text.SpannedString
 import android.text.style.ClickableSpan
+import android.util.AttributeSet
 import android.view.View
-import android.widget.TextView
-import androidx.core.widget.addTextChangedListener
+import android.view.ViewTreeObserver
+import androidx.core.text.buildSpannedString
+import androidx.core.text.inSpans
 import com.adityaamolbavadekar.messenger.R
 
-/**
- * Wraps [TextView] to provide "Read more" facility.
- * */
-class ReadMoreTextView constructor(private val textView: TextView) :
-    View.OnClickListener {
+class ReadMoreClipTextView @JvmOverloads constructor(
+    context: Context,
+    attrs: AttributeSet? = null,
+    defStyle: Int = 0
+) : androidx.appcompat.widget.AppCompatTextView(context, attrs, defStyle) {
 
-    private var longText: SpannableStringBuilder? = null
-    private val readMoreText: String = textView.context.getString(R.string.read_more)
-    private var isExpanded = false
+    private var text: CharSequence? = null
+    private var state: State = State.READ_MORE
+    private var trimSize = DEFAULT_READ_MORE_LINES
+    private var lineEndIndex = 0
+    private val readMoreText = context.getString(R.string.read_more)
 
     init {
-        textView.addTextChangedListener {
-            validate(it?.toString())
+        onGlobalLayoutLineEndIndex()
+        internalSetText()
+    }
+
+    fun isInReadMore() = state == State.READ_MORE
+
+    fun toggleReadMore() {
+        state = when (state) {
+            State.READ_MORE -> State.NORMAL
+            State.NORMAL -> State.READ_MORE
+        }
+        internalSetText()
+    }
+
+    fun setTrimLines(lines: Int) {
+        trimSize = lines
+    }
+
+    override fun setText(text: CharSequence?, type: BufferType?) {
+        this.text = text
+        internalSetText()
+    }
+
+    private fun internalSetText() {
+        super.setText(getSpannedText(), BufferType.SPANNABLE)
+    }
+
+    private fun getSpannedText(): SpannedString {
+        return when (state) {
+            State.READ_MORE -> {
+                if (lineCount > trimSize) addReadMoreFunctionality()
+                else buildSpannedString { append(text) }
+            }
+            State.NORMAL -> buildSpannedString { append(text) }
         }
     }
 
-    private fun validate(text: String?) {
-        if (text == null) return
-        longText = SpannableStringBuilder(text)
-        if (textView.lineCount > MAX_READ_MORE_LINES) addReadMoreFunctionality()
-    }
-
-    private fun addReadMoreFunctionality() {
-        val lineEndIndex = textView.layout.getLineEnd(MAX_READ_MORE_LINES - 1)
-        val baseText = longText!!.subSequence(0..lineEndIndex)
-        val newText = "$baseText $readMoreText"
-        val indexOfReadMoreText = newText.indexOf(readMoreText)
-        val clickable = object : ClickableSpan() {
-            override fun onClick(widget: View) {
-                onClick(widget)
+    private fun addReadMoreFunctionality(): SpannedString {
+        val trimLength = lineEndIndex - (ELLIPSE.length + readMoreText.length)
+        return buildSpannedString {
+            append(text, 0, trimLength)
+            inSpans(ReadMoreSpan()) {
+                append(ELLIPSE + readMoreText)
             }
         }
-        val stringBuilder = SpannableStringBuilder(newText)
-        stringBuilder.setSpan(
-            clickable,
-            indexOfReadMoreText,
-            indexOfReadMoreText + readMoreText.length,
-            Spanned.SPAN_EXCLUSIVE_EXCLUSIVE
-        )
-        textView.text = stringBuilder
     }
+
+    private fun onGlobalLayoutLineEndIndex() {
+        viewTreeObserver.addOnGlobalLayoutListener(onGlobalLayoutListener)
+    }
+
+    private val onGlobalLayoutListener = object : ViewTreeObserver.OnGlobalLayoutListener {
+        override fun onGlobalLayout() {
+            viewTreeObserver.removeOnGlobalLayoutListener(this)
+            refreshLineIndexData()
+            internalSetText()
+        }
+    }
+
+    private fun refreshLineIndexData() {
+        try {
+            if (lineCount >= trimSize) {
+                lineEndIndex = layout.getLineEnd(trimSize - 1)
+            }
+        } catch (_: Exception) {
+        }
+    }
+
+    enum class State { READ_MORE, NORMAL }
 
     companion object {
-        private const val MAX_READ_MORE_LINES = 120
-        fun wrap(textView: TextView) = ReadMoreTextView(textView)
+        private const val DEFAULT_READ_MORE_LINES = 25
+        private const val ELLIPSE = "... "
     }
 
-    override fun onClick(v: View?) {}
+    inner class ReadMoreSpan : ClickableSpan() {
+        override fun onClick(p0: View) = toggleReadMore()
+    }
 
 }
