+++ app/src/main/res/layout/conversation_activity.xml
@@ -67,7 +67,7 @@
                         android:minHeight="72dp"
                         android:orientation="horizontal">
 
-                        <com.adityaamolbavadekar.messenger.views.AvatarImageView
+                        <com.adityaamolbavadekar.messenger.views.image.AvatarImageView
                             android:id="@+id/profile_picture_image_view"
                             android:layout_width="@dimen/toolbar_profile_picture_size"
                             android:layout_height="@dimen/toolbar_profile_picture_size"
