+++ app/src/main/res/layout/conversation_toolbar.xml
@@ -35,7 +35,7 @@
             android:minHeight="?actionBarSize"
             android:orientation="horizontal">
 
-            <com.adityaamolbavadekar.messenger.views.AvatarImageView
+            <com.adityaamolbavadekar.messenger.views.image.AvatarImageView
                 android:id="@+id/profile_picture_image_view"
                 android:layout_width="@dimen/toolbar_profile_picture_size"
                 android:layout_height="@dimen/toolbar_profile_picture_size"
