+++ app/src/main/res/layout/emoji_reaction_item.xml
@@ -7,7 +7,8 @@
     android:layout_height="wrap_content"
     android:layout_margin="2dp"
     android:orientation="horizontal"
-    android:padding="4dp">
+    android:paddingVertical="2dp"
+    android:paddingHorizontal="4dp">
 
     <TextView
         android:id="@+id/emojiTextView"
