--- app/src/main/res/layout/message_reactions_indicator_item.xml
@@ -1,57 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!--
-  ~    Copyright 2023 Aditya Bavadekar
-  ~
-  ~    Licensed under the Apache License, Version 2.0 (the "License");
-  ~    you may not use this file except in compliance with the License.
-  ~    You may obtain a copy of the License at
-  ~
-  ~        http://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~    Unless required by applicable law or agreed to in writing, software
-  ~    distributed under the License is distributed on an "AS IS" BASIS,
-  ~    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  ~    See the License for the specific language governing permissions and
-  ~    limitations under the License.
-  -->
-
-<com.google.android.material.card.MaterialCardView xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/reactionsHolder"
-    style="?materialCardViewFilledStyle"
-    android:layout_width="wrap_content"
-    android:layout_height="wrap_content"
-    android:visibility="gone"
-    android:layout_gravity="start|center_vertical"
-    app:layout_constraintTop_toTopOf="parent"
-    android:layout_marginStart="@dimen/dimen_8"
-    android:layout_marginVertical="@dimen/dimen_4"
-    app:layout_constraintBottom_toBottomOf="parent"
-    app:layout_constraintStart_toStartOf="parent"
-    app:cardCornerRadius="@dimen/message_reactions_indicator_radius"
-    app:cardElevation="0dp"
-    tools:visibility="visible">
-
-    <LinearLayout
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_gravity="center"
-        android:background="?selectableItemBackgroundBorderless"
-        android:padding="@dimen/dimen_4">
-
-        <TextView
-            android:id="@+id/reaction_indicator"
-            android:layout_width="wrap_content"
-            android:layout_height="match_parent"
-            android:paddingHorizontal="@dimen/dimen_2"
-            android:gravity="center"
-            android:text="@string/reacted_count_formatted"
-            tools:text="99+ reacted"
-            android:visibility="gone"
-            tools:visibility="visible" />
-
-        <!--TODO-->
-
-    </LinearLayout>
-</com.google.android.material.card.MaterialCardView>
