+++ app/src/main/res/layout/photo_attachments_view_layout.xml
@@ -18,28 +18,12 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/holder"
-    style="@style/Messenger.Widget.PhotoAttachmentsView.Card"
+    style="@style/Messenger.Widget.PhotoAttachmentsView"
     android:layout_width="@dimen/message_max_width"
     android:maxWidth="@dimen/message_image_height"
     app:cardCornerRadius="@dimen/message_corner_radius"
     app:cardElevation="0dp"
-    app:cardForegroundColor="@android:color/transparent"
-    app:contentPaddingBottom="10dp"
-    app:contentPaddingLeft="10dp"
-    app:contentPaddingRight="10dp"
-    app:contentPaddingTop="10dp">
-
-    <com.google.android.material.card.MaterialCardView
-        style="?materialCardViewFilledStyle"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginVertical="@dimen/dimen_4"
-        android:clickable="false"
-        android:focusable="false"
-        android:maxWidth="@dimen/message_image_height"
-        android:minWidth="@dimen/message_max_width"
-        app:cardCornerRadius="@dimen/message_corner_radius"
-        app:cardForegroundColor="@android:color/transparent">
+    app:cardForegroundColor="@android:color/transparent">
 
         <LinearLayout
             android:layout_width="match_parent"
@@ -142,7 +126,5 @@
             </LinearLayout>
         </LinearLayout>
 
-
-    </com.google.android.material.card.MaterialCardView>
 
 </com.google.android.material.card.MaterialCardView>
